JBPAPP-9018 IBM JDK handles Class.getMethods() in another way than Oracle JDK, it causes problems to Arquillian deployments.
