+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/web/security/form/WebSecurityFORMTestCase.java
@@ -21,28 +21,9 @@
  */
 package org.jboss.as.test.integration.web.security.form;
 
-
-import java.net.URL;
-import java.util.ArrayList;
-import java.util.List;
-
-import org.apache.http.Header;
-import org.apache.http.HttpEntity;
-import org.apache.http.HttpResponse;
-import org.apache.http.NameValuePair;
-import org.apache.http.StatusLine;
-import org.apache.http.client.entity.UrlEncodedFormEntity;
-import org.apache.http.client.methods.HttpGet;
-import org.apache.http.client.methods.HttpPost;
-import org.apache.http.cookie.Cookie;
-import org.apache.http.impl.client.DefaultHttpClient;
-import org.apache.http.message.BasicNameValuePair;
-import org.apache.http.protocol.HTTP;
-import org.apache.http.util.EntityUtils;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.container.test.api.RunAsClient;
 import org.jboss.arquillian.junit.Arquillian;
-import org.jboss.arquillian.test.api.ArquillianResource;
 import org.jboss.as.arquillian.api.ServerSetup;
 import org.jboss.as.test.integration.web.security.SecuredServlet;
 import org.jboss.as.test.integration.web.security.WebSecurityPasswordBasedBase;
@@ -51,17 +32,15 @@ import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
 import org.junit.runner.RunWith;
 
-import static org.junit.Assert.assertEquals;
-
 /**
  * Unit Test web security
- *
+ * 
  * @author Anil Saldhana
  */
 @RunWith(Arquillian.class)
 @RunAsClient
 @ServerSetup(WebTestsSecurityDomainSetup.class)
-public class WebSecurityFORMTestCase extends WebSecurityPasswordBasedBase {
+public class WebSecurityFORMTestCase extends AbstractWebSecurityFORMTestCase {
 
     @Deployment
     public static WebArchive deployment() throws Exception {
@@ -79,89 +58,5 @@ public class WebSecurityFORMTestCase extends WebSecurityPasswordBasedBase {
         war.addAsResource(WebSecurityFORMTestCase.class.getPackage(), "roles.properties", "roles.properties");
         WebSecurityPasswordBasedBase.printWar(war);
         return war;
-    }
-
-    @ArquillianResource
-    private URL url;
-
-    protected void makeCall(String user, String pass, int expectedStatusCode) throws Exception {
-        DefaultHttpClient httpclient = new DefaultHttpClient();
-        try {
-            String req = url.toExternalForm() + "secured/";
-            HttpGet httpget = new HttpGet(req);
-
-            HttpResponse response = httpclient.execute(httpget);
-
-            HttpEntity entity = response.getEntity();
-            if (entity != null)
-                EntityUtils.consume(entity);
-
-            // We should get the Login Page
-            StatusLine statusLine = response.getStatusLine();
-            System.out.println("Login form get: " + statusLine);
-            assertEquals(200, statusLine.getStatusCode());
-
-            System.out.println("Initial set of cookies:");
-            List<Cookie> cookies = httpclient.getCookieStore().getCookies();
-            if (cookies.isEmpty()) {
-                System.out.println("None");
-            } else {
-                for (int i = 0; i < cookies.size(); i++) {
-                    System.out.println("- " + cookies.get(i).toString());
-                }
-            }
-            req = url.toExternalForm() + "secured/j_security_check";
-            // We should now login with the user name and password
-            HttpPost httpPost = new HttpPost(req);
-
-            List<NameValuePair> nvps = new ArrayList<NameValuePair>();
-            nvps.add(new BasicNameValuePair("j_username", user));
-            nvps.add(new BasicNameValuePair("j_password", pass));
-
-            httpPost.setEntity(new UrlEncodedFormEntity(nvps, HTTP.UTF_8));
-
-            response = httpclient.execute(httpPost);
-            entity = response.getEntity();
-            if (entity != null)
-                EntityUtils.consume(entity);
-
-            statusLine = response.getStatusLine();
-
-            // Post authentication - we have a 302
-            assertEquals(302, statusLine.getStatusCode());
-            Header locationHeader = response.getFirstHeader("Location");
-            String location = locationHeader.getValue();
-
-            HttpGet httpGet = new HttpGet(location);
-            response = httpclient.execute(httpGet);
-
-            entity = response.getEntity();
-            if (entity != null)
-                EntityUtils.consume(entity);
-
-            System.out.println("Post logon cookies:");
-            cookies = httpclient.getCookieStore().getCookies();
-            if (cookies.isEmpty()) {
-                System.out.println("None");
-            } else {
-                for (int i = 0; i < cookies.size(); i++) {
-                    System.out.println("- " + cookies.get(i).toString());
-                }
-            }
-
-            // Either the authentication passed or failed based on the expected status code
-            statusLine = response.getStatusLine();
-            assertEquals(expectedStatusCode, statusLine.getStatusCode());
-        } finally {
-            // When HttpClient instance is no longer needed,
-            // shut down the connection manager to ensure
-            // immediate deallocation of all system resources
-            httpclient.getConnectionManager().shutdown();
-        }
-    }
-
-    @Override
-    public String getContextPath() {
-        return "web-secure";
     }
 }
