+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/web/security/form/WebSecurityJBossWebXmlSecurityRolesTestCase.java
@@ -21,13 +21,9 @@
  */
 package org.jboss.as.test.integration.web.security.form;
 
-import java.net.URL;
-
 import org.jboss.arquillian.container.test.api.Deployment;
-import org.jboss.arquillian.container.test.api.OperateOnDeployment;
 import org.jboss.arquillian.container.test.api.RunAsClient;
 import org.jboss.arquillian.junit.Arquillian;
-import org.jboss.arquillian.test.api.ArquillianResource;
 import org.jboss.as.arquillian.api.ServerSetup;
 import org.jboss.as.test.integration.web.security.SecuredServlet;
 import org.jboss.as.test.integration.web.security.WebSecurityPasswordBasedBase;
@@ -39,22 +35,15 @@ import org.junit.runner.RunWith;
 
 /**
  * Unit Test web security
- *
+ * 
  * @author <a href="mailto:mmoyses@redhat.com">Marcus Moyses</a>
  */
 @RunWith(Arquillian.class)
 @RunAsClient
 @ServerSetup(WebTestsSecurityDomainSetup.class)
-public class WebSecurityJBossWebXmlSecurityRolesTestCase extends WebSecurityFORMTestCase {
+public class WebSecurityJBossWebXmlSecurityRolesTestCase extends AbstractWebSecurityFORMTestCase {
 
-    public static final String deploymentName = "web-secure.war";
-
-    @ArquillianResource
-
-    @OperateOnDeployment(deploymentName)
-    URL deploymentUrl;
-
-    @Deployment(name = deploymentName, order = 1, testable = false)
+    @Deployment(testable = false)
     public static WebArchive deployment() throws Exception {
 
         WebArchive war = ShrinkWrap.create(WebArchive.class, "web-secure.war");
@@ -63,18 +52,20 @@ public class WebSecurityJBossWebXmlSecurityRolesTestCase extends WebSecurityFORM
         war.addAsWebResource(WebSecurityJBossWebXmlSecurityRolesTestCase.class.getPackage(), "login.jsp", "login.jsp");
         war.addAsWebResource(WebSecurityJBossWebXmlSecurityRolesTestCase.class.getPackage(), "error.jsp", "error.jsp");
 
-        war.addAsWebInfResource(WebSecurityJBossWebXmlSecurityRolesTestCase.class.getPackage(), "jboss-web-role-mapping.xml", "jboss-web.xml");
+        war.addAsWebInfResource(WebSecurityJBossWebXmlSecurityRolesTestCase.class.getPackage(), "jboss-web-role-mapping.xml",
+                "jboss-web.xml");
         war.addAsWebInfResource(WebSecurityJBossWebXmlSecurityRolesTestCase.class.getPackage(), "web.xml", "web.xml");
 
-        war.addAsResource(WebSecurityJBossWebXmlSecurityRolesTestCase.class.getPackage(), "users.properties", "users.properties");
-        war.addAsResource(WebSecurityJBossWebXmlSecurityRolesTestCase.class.getPackage(), "roles.properties", "roles.properties");
+        war.addAsResource(WebSecurityJBossWebXmlSecurityRolesTestCase.class.getPackage(), "users.properties",
+                "users.properties");
+        war.addAsResource(WebSecurityJBossWebXmlSecurityRolesTestCase.class.getPackage(), "roles.properties",
+                "roles.properties");
         WebSecurityPasswordBasedBase.printWar(war);
         return war;
     }
 
-
     /**
-     * Override test behavior to check if role mapping in jboss-web.xml works.
+     * Negative test as marcus doesn't have proper role in module mapping created.
      */
     @Override
     @Test
@@ -84,7 +75,7 @@ public class WebSecurityJBossWebXmlSecurityRolesTestCase extends WebSecurityFORM
 
     /**
      * Negative test to see if mapping is not performed on username instead of role.
-     *
+     * 
      * @throws Exception
      */
     @Test
